+++ NEWS.md
@@ -2,7 +2,7 @@
 
 ## CytoPipelineUtils 0.99.1
 
-- added support for FlowJo manual gating as a CytiProcessingStep
+- added support for FlowJo manual gating as a CytoProcessingStep
 - used file.path() instead of paste0() to construct file paths 
 
 ## CytoPipelineUtils 0.99.0